Clarify Node types package validation
